+++ lib/savon/soap/fault.rb
@@ -30,7 +30,7 @@ module Savon
 
       # Returns the SOAP response body as a Hash.
       def to_hash
-        @hash ||= Savon::SOAP::XML.to_hash http.body
+        @hash ||= Nori.parse(http.body)[:envelope][:body]
       end
 
     private