+++ spec/savon/soap/response_spec.rb
@@ -116,44 +116,95 @@ describe Savon::SOAP::Response do
     end
   end
 
-  describe "#header" do
-    it "should return the SOAP response header as a Hash" do
-      response = soap_response :body => Fixture.response(:header)
-      response.header.should include(:session_number => "ABCD1234")
-    end
-  end
-
   describe "#[]" do
     it "should return the SOAP response body as a Hash" do
       soap_response[:authenticate_response][:return].should ==
         Fixture.response_hash(:authentication)[:authenticate_response][:return]
     end
+
+    it "should throw an exception when the response body isn't parsable" do
+      lambda { invalid_soap_response.body }.should raise_error Savon::SOAP::InvalidResponseError
+    end
   end
 
-  describe "#to_hash" do
-    it "should return the SOAP response body as a Hash" do
-      soap_response.to_hash[:authenticate_response][:return].should ==
-        Fixture.response_hash(:authentication)[:authenticate_response][:return]
+  describe "#header" do
+    it "should return the SOAP response header as a Hash" do
+      response = soap_response :body => Fixture.response(:header)
+      response.header.should include(:session_number => "ABCD1234")
+    end
+
+    it "should throw an exception when the response header isn't parsable" do
+      lambda { invalid_soap_response.header }.should raise_error Savon::SOAP::InvalidResponseError
+    end
+  end
+
+  %w(body to_hash).each do |method|
+    describe "##{method}" do
+      it "should return the SOAP response body as a Hash" do
+        soap_response.send(method)[:authenticate_response][:return].should ==
+          Fixture.response_hash(:authentication)[:authenticate_response][:return]
+      end
+
+      it "should return a Hash for a SOAP multiRef response" do
+        hash = soap_response(:body => Fixture.response(:multi_ref)).send(method)
+
+        hash[:list_response].should be_a(Hash)
+        hash[:multi_ref].should be_an(Array)
+      end
+
+      it "should add existing namespaced elements as an array" do
+        hash = soap_response(:body => Fixture.response(:list)).send(method)
+
+        hash[:multi_namespaced_entry_response][:history].should be_a(Hash)
+        hash[:multi_namespaced_entry_response][:history][:case].should be_an(Array)
+      end
     end
   end
 
   describe "#to_array" do
-    it "should delegate to Savon::SOAP::XML.to_array" do
-      Savon::SOAP::XML.expects(:to_array).with(soap_response.to_hash, :authenticate_response, :return)
-      soap_response.to_array :authenticate_response, :return
+    context "when the given path exists" do
+      it "should return an Array containing the path value" do
+        soap_response.to_array(:authenticate_response, :return).should ==
+          [Fixture.response_hash(:authentication)[:authenticate_response][:return]]
+      end
+    end
+
+    context "when the given path returns nil" do
+      it "should return an empty Array" do
+        soap_response.to_array(:authenticate_response, :undefined).should == []
+      end
+    end
+
+    context "when the given path does not exist at all" do
+      it "should return an empty Array" do
+        soap_response.to_array(:authenticate_response, :some, :undefined, :path).should == []
+      end
     end
   end
 
-  describe "#basic_hash" do
+  describe "#hash" do
     it "should return the complete SOAP response XML as a Hash" do
       response = soap_response :body => Fixture.response(:header)
-      response.basic_hash["soap:Envelope"]["soap:Header"]["SessionNumber"].should == "ABCD1234"
+      response.hash[:envelope][:header][:session_number].should == "ABCD1234"
     end
   end
 
   describe "#to_xml" do
     it "should return the raw SOAP response body" do
       soap_response.to_xml.should == Fixture.response(:authentication)
+    end
+  end
+
+  describe "#doc" do
+    it "returns a Nokogiri::XML::Document for the SOAP response XML" do
+      soap_response.doc.should be_a(Nokogiri::XML::Document)
+    end
+  end
+
+  describe "#xpath" do
+    it "permits XPath access to elements in the request" do
+      soap_response.xpath("//client").first.inner_text.should == "radclient"
+      soap_response.xpath("//ns2:authenticateResponse/return/success").first.inner_text.should == "true"
     end
   end
 
@@ -198,6 +249,13 @@ describe Savon::SOAP::Response do
 
   def http_error_response
     soap_response :code => 404, :body => "Not found"
+  end
+
+  def invalid_soap_response(options={})
+    defaults = { :code => 200, :headers => {}, :body => "I'm not SOAP" } 
+    response = defaults.merge options
+
+    Savon::SOAP::Response.new HTTPI::Response.new(response[:code], response[:headers], response[:body])
   end
 
 end